amazon-fba-image-optimizer/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.0.1
opencv-python==4.8.1.78
numpy==1.24.3
boto3==1.29.7
python-dotenv==1.0.0
pydantic==2.4.2